requirements.txt
Flask~=3.0.3
spacy~=3.7.4
Werkzeug~=3.0.3
SQLAlchemy~=2.0.30
pyodbc~=5.1.0
alembic~=1.13.1
Flask-Bcrypt~=1.0.1
WTForms~=3.1.2
pytest~=8.2.0
textblob~=0.17.1
docx2txt~=0.8
pdfminer.six~=20240706
nltk~=3.6.7
